requirements.txt
curl-cffi
beautifulsoup4
google-generativeai
python-dotenv
pydantic
instructor